Hoist test file parsing logic
(cherry picked from commit 48e9b3c4759b5ae2b6029ea95b59b604bdc2f5e0)
