+++ compiler-plugin/src/test/kotlin/com/google/devtools/ksp/test/AbstractKSPCompilerPluginTest.kt
@@ -19,11 +19,15 @@ import org.jetbrains.kotlin.test.model.TestModule
 import org.jetbrains.kotlin.test.services.TestServices
 import org.jetbrains.kotlin.test.services.compilerConfigurationProvider
 import org.jetbrains.kotlin.test.services.javaFiles
-import org.junit.jupiter.api.Assertions
 import java.io.File
 
 abstract class AbstractKSPCompilerPluginTest : AbstractKSPTest(FrontendKinds.ClassicFrontend) {
-    override fun runTest(testServices: TestServices, mainModule: TestModule, libModules: List<TestModule>) {
+    override fun runTest(
+        testServices: TestServices,
+        mainModule: TestModule,
+        libModules: List<TestModule>,
+        testProcessor: AbstractTestProcessor,
+    ): List<String> {
         val compilerConfiguration = testServices.compilerConfigurationProvider.getCompilerConfiguration(mainModule)
         compilerConfiguration.put(CommonConfigurationKeys.MODULE_NAME, mainModule.name)
         compilerConfiguration.put(CommonConfigurationKeys.LOOKUP_TRACKER, DualLookupTracker())
@@ -47,16 +51,6 @@ abstract class AbstractKSPCompilerPluginTest : AbstractKSPTest(FrontendKinds.Cla
             false
         )
 
-        val contents = mainModule.files.first().originalFile.readLines()
-        val testProcessorName = contents
-            .filter { it.startsWith(TEST_PROCESSOR) }
-            .single()
-            .substringAfter(TEST_PROCESSOR)
-            .trim()
-        val testProcessor: AbstractTestProcessor =
-            Class.forName("com.google.devtools.ksp.processor.$testProcessorName")
-                .getDeclaredConstructor().newInstance() as AbstractTestProcessor
-
         val testRoot = mainModule.testRoot
         val analysisExtension =
             KotlinSymbolProcessingExtension(
@@ -78,12 +72,6 @@ abstract class AbstractKSPCompilerPluginTest : AbstractKSPTest(FrontendKinds.Cla
 
         GenerationUtils.compileFilesTo(ktFiles, kotlinCoreEnvironment, mainModule.outDir)
 
-        val result = testProcessor.toResult()
-        val expectedResults = contents
-            .dropWhile { !it.startsWith(EXPECTED_RESULTS) }
-            .drop(1)
-            .takeWhile { !it.startsWith("// END") }
-            .map { it.substring(3).trim() }
-        Assertions.assertEquals(expectedResults.joinToString("\n"), result.joinToString("\n"))
+        return testProcessor.toResult()
     }
 }
