+++ kotlin-analysis-api/src/test/kotlin/com/google/devtools/ksp/impl/test/AbstractKSPAATest.kt
@@ -39,7 +39,6 @@ import org.jetbrains.kotlin.test.services.TestServices
 import org.jetbrains.kotlin.test.services.compilerConfigurationProvider
 import org.jetbrains.kotlin.test.services.isKtFile
 import org.jetbrains.kotlin.test.services.javaFiles
-import org.junit.jupiter.api.Assertions
 import java.io.File
 
 abstract class AbstractKSPAATest : AbstractKSPTest(FrontendKinds.FIR) {
@@ -53,7 +52,12 @@ abstract class AbstractKSPAATest : AbstractKSPTest(FrontendKinds.FIR) {
         }
     }
 
-    override fun runTest(testServices: TestServices, mainModule: TestModule, libModules: List<TestModule>) {
+    override fun runTest(
+        testServices: TestServices,
+        mainModule: TestModule,
+        libModules: List<TestModule>,
+        testProcessor: AbstractTestProcessor
+    ): List<String> {
         val compilerConfiguration = testServices.compilerConfigurationProvider.getCompilerConfiguration(mainModule)
         compilerConfiguration.put(CommonConfigurationKeys.MODULE_NAME, mainModule.name)
         compilerConfiguration.put(CommonConfigurationKeys.LOOKUP_TRACKER, DualLookupTracker())
@@ -90,25 +94,9 @@ abstract class AbstractKSPAATest : AbstractKSPTest(FrontendKinds.FIR) {
             kotlinCoreEnvironment.projectEnvironment.environment.jarFileSystem as CoreJarFileSystem
         )
 
-        val contents = mainModule.files.first().originalFile.readLines()
-        val testProcessorName = contents
-            .filter { it.startsWith(TEST_PROCESSOR) }
-            .single()
-            .substringAfter(TEST_PROCESSOR)
-            .trim()
-        val testProcessor: AbstractTestProcessor =
-            Class.forName("com.google.devtools.ksp.processor.$testProcessorName")
-                .getDeclaredConstructor().newInstance() as AbstractTestProcessor
-
         val resolver = ResolverAAImpl(ktFiles)
         testProcessor.process(resolver)
 
-        val result = testProcessor.toResult()
-        val expectedResults = contents
-            .dropWhile { !it.startsWith(EXPECTED_RESULTS) }
-            .drop(1)
-            .takeWhile { !it.startsWith("// END") }
-            .map { it.substring(3).trim() }
-        Assertions.assertEquals(expectedResults.joinToString("\n"), result.joinToString("\n"))
+        return testProcessor.toResult()
     }
 }
